+++ lib/tilt/template.rb
@@ -99,11 +99,7 @@ module Tilt
     # block is given, it is typically available within the template via
     # +yield+.
     def render(scope=nil, locals=nil, &block)
-      current_template = Thread.current[:tilt_current_template]
-      Thread.current[:tilt_current_template] = self
       evaluate(scope || Object.new, locals || EMPTY_HASH, &block)
-    ensure
-      Thread.current[:tilt_current_template] = current_template
     end
 
     # The basename of the template file.