+++ test/tilt_template_test.rb
@@ -353,23 +353,6 @@ describe "tilt/template" do
     assert_equal "Hey Bob!", inst.render(_BasicPerson)
   end
 
-  deprecated "populates Tilt.current_template during rendering" do
-    inst = _SourceGeneratingMockTemplate.new { '#{$inst = Tilt.current_template}' }
-    inst.render
-    assert_equal inst, $inst
-    assert_nil Tilt.current_template
-  end
-
-  deprecated "populates Tilt.current_template in nested rendering" do
-    inst1 = _SourceGeneratingMockTemplate.new { '#{$inst.render; $inst1 = Tilt.current_template}' }
-    inst2 = _SourceGeneratingMockTemplate.new { '#{$inst2 = Tilt.current_template}' }
-    $inst = inst2
-    inst1.render
-    assert_equal inst1, $inst1
-    assert_equal inst2, $inst2
-    assert_nil Tilt.current_template
-  end
-
   if RUBY_VERSION >= '2.3'
     _FrozenStringMockTemplate = Class.new(_PreparingMockTemplate) do
       def freeze_string_literals?